chore: use agreed up Configuration class suffix Config
https://github.com/dhis2/wow-backend/blob/master/guides/spring_conventions.md\#spring-configuration
