+++ dhis-2/dhis-support/dhis-support-cache-invalidation/src/test/java/org/hisp/dhis/cacheinvalidation/redis/CacheInvalidationListenerTest.java
@@ -108,7 +108,7 @@ class CacheInvalidationListenerTest {
             + "ROLE"
             + ":"
             + "1";
-    cacheInvalidationListener.message(CacheInvalidationConfiguration.CHANNEL_NAME, message);
+    cacheInvalidationListener.message(CacheInvalidationConfig.CHANNEL_NAME, message);
 
     verify(queryCacheManager, times(0)).evictQueryCache(any(), any());
     verify(sessionFactory.getCache(), times(1)).evictCollectionData(any(), any());
@@ -119,7 +119,7 @@ class CacheInvalidationListenerTest {
   @DisplayName("Should not call evict cache on INSERT messages")
   void testInsertMessage() {
     String message = "SERVER_B" + ":" + "INSERT" + ":" + "org.hisp.dhis.user.User" + ":" + "1";
-    cacheInvalidationListener.message(CacheInvalidationConfiguration.CHANNEL_NAME, message);
+    cacheInvalidationListener.message(CacheInvalidationConfig.CHANNEL_NAME, message);
 
     verify(queryCacheManager, times(1)).evictQueryCache(any(), any());
     verify(sessionFactory.getCache(), times(0)).evict(any(), any());
@@ -130,7 +130,7 @@ class CacheInvalidationListenerTest {
   @DisplayName("Should call evict cache on UPDATE messages")
   void testUpdateMessage() {
     String message = "SERVER_B" + ":" + "UPDATE" + ":" + "org.hisp.dhis.user.User" + ":" + "1";
-    cacheInvalidationListener.message(CacheInvalidationConfiguration.CHANNEL_NAME, message);
+    cacheInvalidationListener.message(CacheInvalidationConfig.CHANNEL_NAME, message);
 
     verify(sessionFactory.getCache(), times(1)).evict(any(), any());
   }
@@ -139,7 +139,7 @@ class CacheInvalidationListenerTest {
   @DisplayName("Should call evict cache on DELETE messages")
   void testDeleteMessage() {
     String message = "SERVER_B" + ":" + "DELETE" + ":" + "org.hisp.dhis.user.User" + ":" + "1";
-    cacheInvalidationListener.message(CacheInvalidationConfiguration.CHANNEL_NAME, message);
+    cacheInvalidationListener.message(CacheInvalidationConfig.CHANNEL_NAME, message);
 
     verify(queryCacheManager, times(1)).evictQueryCache(any(), any());
     verify(sessionFactory.getCache(), times(1)).evict(any(), any());
