+++ dhis-2/dhis-support/dhis-support-system/src/main/java/org/hisp/dhis/leader/election/LeaderElectionConfig.java
@@ -25,35 +25,38 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.webapi.filter;
+package org.hisp.dhis.leader.election;
 
-import javax.servlet.Filter;
 import org.hisp.dhis.condition.RedisDisabledCondition;
+import org.hisp.dhis.condition.RedisEnabledCondition;
+import org.hisp.dhis.external.conf.DhisConfigurationProvider;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Conditional;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.web.filter.CharacterEncodingFilter;
+import org.springframework.data.redis.core.StringRedisTemplate;
 
 /**
- * Configuration registered if {@link RedisDisabledCondition} matches to true. This serves as a
- * fallback to spring-session if redis is disabled. Since web.xml has a
- * "springSessionRepositoryFilter" mapped to all urls, the container will expect a filter bean with
- * that name. Therefore we define a dummy {@link Filter} named springSessionRepositoryFilter. Here
- * we define a {@link CharacterEncodingFilter} without setting any encoding so that requests will
- * simply pass through the filter.
+ * Configures leaderManager that takes care of node leader elections.
  *
  * @author Ameen Mohamed
  */
 @Configuration
-@Conditional(RedisDisabledCondition.class)
-public class DefaultSessionConfiguration {
-  /**
-   * Defines a {@link CharacterEncodingFilter} named springSessionRepositoryFilter
-   *
-   * @return a {@link CharacterEncodingFilter} without specifying encoding.
-   */
-  @Bean
-  public Filter springSessionRepositoryFilter() {
-    return new CharacterEncodingFilter();
+public class LeaderElectionConfig {
+  @Autowired private DhisConfigurationProvider dhisConfigurationProvider;
+
+  @Bean(name = "leaderManager")
+  @Conditional(RedisEnabledCondition.class)
+  public LeaderManager redisLeaderManager(
+      @Autowired(required = false) @Qualifier("stringRedisTemplate")
+          StringRedisTemplate stringRedisTemplate) {
+    return new RedisLeaderManager(stringRedisTemplate, dhisConfigurationProvider);
+  }
+
+  @Bean(name = "leaderManager")
+  @Conditional(RedisDisabledCondition.class)
+  public LeaderManager noOpLeaderManager() {
+    return new NoOpLeaderManager(dhisConfigurationProvider);
   }
 }
