+++ dhis-2/dhis-support/dhis-support-test/src/main/java/org/hisp/dhis/test/config/PostgresDhisTestConfig.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2004-2023, University of Oslo
+ * Copyright (c) 2004-2022, University of Oslo
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -27,39 +27,21 @@
  */
 package org.hisp.dhis.test.config;
 
+import org.hisp.dhis.external.conf.DhisConfigurationProvider;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.core.session.SessionRegistry;
-import org.springframework.security.core.session.SessionRegistryImpl;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.ldap.authentication.LdapAuthenticator;
-import org.springframework.security.ldap.userdetails.LdapAuthoritiesPopulator;
+import org.springframework.context.annotation.Profile;
 
 /**
- * @author Morten Svanæs <msvanaes@dhis2.org>
+ * Use this Spring configuration for tests relying on the Postgres DB running in a Docker container.
+ *
+ * @author Gintare Vilkelyte <vilkelyte.gintare@gmail.com>
  */
+@Profile("test-postgres")
 @Configuration
-@ComponentScan("org.hisp.dhis")
-public class IntegrationBaseConfiguration {
+public class PostgresDhisTestConfig {
   @Bean
-  public static SessionRegistry sessionRegistry() {
-    return new SessionRegistryImpl();
-  }
-
-  @Bean
-  public LdapAuthenticator ldapAuthenticator() {
-    return authentication -> null;
-  }
-
-  @Bean
-  public LdapAuthoritiesPopulator ldapAuthoritiesPopulator() {
-    return (dirContextOperations, s) -> null;
-  }
-
-  @Bean
-  public PasswordEncoder encoder() {
-    return new BCryptPasswordEncoder();
+  public DhisConfigurationProvider dhisConfigurationProvider() {
+    return new PostgresDhisConfigurationProvider();
   }
 }
