+++ dhis-2/dhis-test-web-api/src/test/java/org/hisp/dhis/test/webapi/AuthenticationApiTestBase.java
@@ -27,7 +27,8 @@
  */
 package org.hisp.dhis.test.webapi;
 
-import org.hisp.dhis.test.config.H2DhisConfiguration;
+import org.hisp.dhis.test.config.H2DhisTestConfig;
+import org.hisp.dhis.test.webapi.AuthenticationApiTestBase.AuthConfigProviderConfig;
 import org.hisp.dhis.webapi.security.config.WebMvcConfig;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -53,13 +54,13 @@ import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 @ContextConfiguration(
     inheritLocations = false,
     classes = {
-      H2DhisConfiguration.class,
+      H2DhisTestConfig.class,
       WebMvcConfig.class,
-      AuthenticationApiTestBase.AuthConfigProviderConfiguration.class,
+      AuthConfigProviderConfig.class,
     })
 public abstract class AuthenticationApiTestBase extends H2ControllerIntegrationTestBase {
 
-  static class AuthConfigProviderConfiguration {
+  static class AuthConfigProviderConfig {
     @Bean
     public WebSecurityCustomizer webSecurityCustomizer() {
       return web ->
