+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/security/config/WebMvcMetricsConfig.java
@@ -71,7 +71,7 @@ public class WebMvcMetricsConfig {
 
   @Configuration
   @Conditional(WebMvcMetricsDisabledCondition.class)
-  static class DataSourcePoolMetadataMetricsConfiguration {
+  static class DataSourcePoolMetadataMetricsConfig {
     @Bean
     public PassThroughWebMvcMetricsFilter webMetricsFilter() {
       return new PassThroughWebMvcMetricsFilter();
